Merge test

+++ src/graysvr/CObjBase.h
@@ -309,7 +309,7 @@ public:
 	void inline SetNamePool_Fail( TCHAR * ppTitles );
 	bool SetNamePool( LPCTSTR pszName );
 
-	void Sound( SOUND_TYPE id, int iRepeat = 1 ) const; // Play sound effect from this location.
+	void Sound(SOUND_TYPE id, BYTE iRepeat = 1) const; // Play sound effect from this location.
 	void Effect(EFFECT_TYPE motion, ITEMID_TYPE id, const CObjBase * pSource = NULL, BYTE bspeedseconds = 5, BYTE bloop = 1, bool fexplode = false, DWORD color = 0, DWORD render = 0, WORD effectid = 0, WORD explodeid = 0, WORD explodesound = 0, DWORD effectuid = 0, BYTE type = 0) const;
 
 	void r_WriteSafe( CScript & s );
@@ -904,7 +904,7 @@ public:
 		struct
 		{
 			DWORD	m_Sound;	// more1 = SOUND_TYPE
-			int		m_Repeat;	// more2 =
+			BYTE	m_Repeat;	// more2 =
 		} m_itSound;
 
 		// IT_STONE_GUILD
@@ -1764,13 +1764,13 @@ public:
 	};
 
 private:
-	typedef std::vector<Component*> ComponentsContainer;
+	typedef std::vector<Component *> ComponentsContainer;
 	struct DesignDetails
 	{
-		int m_iRevision;
+		DWORD m_iRevision;
 		ComponentsContainer m_vectorComponents;
-		PacketHouseDesign* m_pData;
-		int m_iDataRevision;
+		PacketHouseDesign *m_pData;
+		DWORD m_iDataRevision;
 	};
 	
 	class CGrayMultiCustom : public CGrayMulti
@@ -1844,7 +1844,7 @@ public:
 	const CGRect GetDesignArea();
 	size_t GetFixtureCount(DesignDetails * pDesign = NULL);
 	size_t GetComponentsAt(signed short dx, signed short dy, signed char dz, Component ** pComponents, DesignDetails * pDesign = NULL);
-	int GetRevision(const CClient * pClientSrc = NULL) const;
+	DWORD GetRevision(const CClient * pClientSrc = NULL) const;
 	BYTE GetLevelCount();
 	WORD GetStairID();
